pyproject.toml
[project]
name = "anyimg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.39.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "rich>=14.1.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.405",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.13.2",
]

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.11"
include = ["src", "tests"]
exclude = [".venv", "**/__pycache__"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
